requirements.txt
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.11
idna==3.3
numpy==1.22.2
pandas==1.4.1
pycodestyle==2.8.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
six==1.16.0
surveymonkey-python==0.1.4
toml==0.10.2
typeform==1.1.0
urllib3==1.26.8
